* Move ingestion commandline into splitgraph.commandline (avoid circular imports)
