+++ test/splitgraph/ingestion/test_commandline.py
@@ -3,7 +3,7 @@ from datetime import datetime as dt
 
 from click.testing import CliRunner
 
-from splitgraph.ingestion import csv_import, csv_export
+from splitgraph.commandline.ingestion import csv_import, csv_export
 from test.splitgraph.conftest import INGESTION_RESOURCES
